added user entity, design changes for all of the pages, access control (still needs some changes), routing logic for each role after logging in

+++ src/Controller/ConsumerController.php
@@ -3,11 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\Consumer;
-use App\Form\BusinessTypeFormType;
 use App\Form\ConsumerFormType;
 use App\Repository\ConsumerRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -27,6 +27,8 @@ final class ConsumerController extends AbstractController
     #[Route('/consumer/new', name: 'app_consumer_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $consumer = new Consumer();
 
         $form = $this->createForm(ConsumerFormType::class, $consumer);
@@ -44,8 +46,14 @@ final class ConsumerController extends AbstractController
     }
 
     #[Route('/consumer/{id}/edit', name: 'app_consumer_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request,Consumer $consumer, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Consumer $consumer, EntityManagerInterface $entityManager, Security $security): Response
     {
+        $user = $security->getUser();
+
+        if ((!$user || !$user->getConsumer() || $user->getConsumer()->getId() !== $consumer->getId())) {
+            throw $this->createAccessDeniedException();
+        }
+
         $form = $this->createForm(ConsumerFormType::class, $consumer);
         $form->handleRequest($request);
 
@@ -60,17 +68,32 @@ final class ConsumerController extends AbstractController
     }
 
     #[Route('/consumer/{id}', name: 'app_consumer_view', methods: ['GET'])]
-    public function view(int $id, ConsumerRepository $consumerRepository): Response
+    public function view(int $id, ConsumerRepository $consumerRepository, Security $security): Response
     {
         $consumer = $consumerRepository->find($id);
 
-        return $this->render('consumer/view.html.twig', [
-            'consumer' => $consumer,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('consumer/view.html.twig', [
+                'consumer' => $consumer,
+            ]);
+        }
+
+        $user = $security->getUser();
+
+        if ($user && $user->getConsumer() && $user->getConsumer()->getId() === $consumer->getId()) {
+            return $this->render('consumer/view.html.twig', [
+                'consumer' => $consumer,
+            ]);
+        }
+
+        throw $this->createAccessDeniedException();
     }
+
     #[Route('/consumer/delete/{id}', name: 'app_consumer_delete', methods: ['GET'])]
     public function delete(Consumer $consumer, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $entityManager->remove($consumer);
         $entityManager->flush();
         return $this->redirectToRoute('app_consumer');
